Move bw line class to v3bwfile.py
so that it close to the header and we can reuse methods/functions
