+++ sbws/core/generate.py
@@ -1,5 +1,5 @@
 from sbws.globals import (fail_hard, is_initted)
-from sbws.lib.v3bwfile import V3BwHeader
+from sbws.lib.v3bwfile import V3BwHeader, V3BWLine
 from sbws.lib.resultdump import ResultSuccess
 from sbws.lib.resultdump import load_recent_results_in_datadir
 from sbws.util.timestamp import unixts_to_isodt_str
@@ -9,26 +9,6 @@ import os
 import logging
 
 log = logging.getLogger(__name__)
-
-
-# FIXME: move this to v3bwfile?
-class V3BWLine:
-    # TODO: docstrings
-    def __init__(self, fp, bw, nick, rtts, last_time):
-        # TODO: asserts checking arg types
-        self.fp = fp
-        self.nick = nick
-        # convert to KiB and make sure the answer is at least 1
-        self.bw = max(round(bw / 1024), 1)
-        # convert to ms
-        rtts = [round(r * 1000) for r in rtts]
-        self.rtt = round(median(rtts))
-        self.time = unixts_to_isodt_str(last_time)
-
-    def __str__(self):
-        frmt = 'node_id=${fp} bw={sp} nick={n} rtt={rtt} time={t}'
-        return frmt.format(fp=self.fp, sp=self.bw, n=self.nick, rtt=self.rtt,
-                           t=self.time)
 
 
 def result_data_to_v3bw_line(data, fingerprint):
